+++ netwerk/cache2/CacheIndex.cpp
@@ -223,7 +223,7 @@ NS_IMETHODIMP FileOpenHelper::OnFileOpened(CacheFileHandle *aHandle,
 NS_IMPL_ISUPPORTS(FileOpenHelper, CacheFileIOListener);
 
 
-CacheIndex * CacheIndex::gInstance = nullptr;
+StaticRefPtr<CacheIndex> CacheIndex::gInstance;
 StaticMutex  CacheIndex::sLock;
 
 
@@ -250,8 +250,10 @@ CacheIndex::CacheIndex()
   , mRWBuf(nullptr)
   , mRWBufSize(0)
   , mRWBufPos(0)
+  , mRWPending(false)
   , mJournalReadSuccessfully(false)
   , mFrecencyArraySorted(false)
+  , mAsyncGetDiskConsumptionBlocked(false)
 {
   sLock.AssertCurrentThreadOwns();
   LOG(("CacheIndex::CacheIndex [this=%p]", this));
@@ -261,6 +263,7 @@ CacheIndex::CacheIndex()
 
 CacheIndex::~CacheIndex()
 {
+  sLock.AssertCurrentThreadOwns();
   LOG(("CacheIndex::~CacheIndex [this=%p]", this));
   MOZ_COUNT_DTOR(CacheIndex);
 
@@ -286,7 +289,7 @@ CacheIndex::Init(nsIFile *aCacheDirectory)
   nsresult rv = idx->InitInternal(aCacheDirectory);
   NS_ENSURE_SUCCESS(rv, rv);
 
-  idx.swap(gInstance);
+  gInstance = idx.forget();
   return NS_OK;
 }
 
@@ -313,7 +316,7 @@ CacheIndex::PreShutdown()
 
   StaticMutexAutoLock lock(sLock);
 
-  LOG(("CacheIndex::PreShutdown() [gInstance=%p]", gInstance));
+  LOG(("CacheIndex::PreShutdown() [gInstance=%p]", gInstance.get()));
 
   nsresult rv;
   RefPtr<CacheIndex> index = gInstance;
@@ -345,17 +348,14 @@ CacheIndex::PreShutdown()
   }
 
   nsCOMPtr<nsIRunnable> event;
-  event = NS_NewRunnableMethod(index, &CacheIndex::PreShutdownInternal);
+  event = NewRunnableMethod(index, &CacheIndex::PreShutdownInternal);
 
-  RefPtr<CacheIOThread> ioThread = CacheFileIOManager::IOThread();
-  MOZ_ASSERT(ioThread);
+  nsCOMPtr<nsIEventTarget> ioTarget = CacheFileIOManager::IOTarget();
+  MOZ_ASSERT(ioTarget);
 
-  // Executing PreShutdownInternal() on WRITE level ensures that read/write
-  // events holding pointer to mRWBuf will be executed before we release the
-  // buffer by calling FinishRead()/FinishWrite() in PreShutdownInternal(), but
-  // it will be executed before any queued event on INDEX level. That's OK since
-  // we don't want to wait until updating of the index finishes.
-  rv = ioThread->Dispatch(event, CacheIOThread::WRITE);
+  // PreShutdownInternal() will be executed before any queued event on INDEX
+  // level. That's OK since we don't want to wait for any operation in progess.
+  rv = ioTarget->Dispatch(event, nsIEventTarget::DISPATCH_NORMAL);
   if (NS_FAILED(rv)) {
     NS_WARNING("CacheIndex::PreShutdown() - Can't dispatch event");
     LOG(("CacheIndex::PreShutdown() - Can't dispatch event" ));
@@ -410,10 +410,9 @@ CacheIndex::Shutdown()
 
   StaticMutexAutoLock lock(sLock);
 
-  LOG(("CacheIndex::Shutdown() [gInstance=%p]", gInstance));
+  LOG(("CacheIndex::Shutdown() [gInstance=%p]", gInstance.get()));
 
-  RefPtr<CacheIndex> index;
-  index.swap(gInstance);
+  RefPtr<CacheIndex> index = gInstance.forget();
 
   if (!index) {
     return NS_ERROR_NOT_INITIALIZED;
@@ -442,10 +441,10 @@ CacheIndex::Shutdown()
     case READY:
       if (index->mIndexOnDiskIsValid && !index->mDontMarkIndexClean) {
         if (!sanitize && NS_FAILED(index->WriteLogToDisk())) {
-          index->RemoveIndexFromDisk();
+          index->RemoveJournalAndTempFile();
         }
       } else {
-        index->RemoveIndexFromDisk();
+        index->RemoveJournalAndTempFile();
       }
       break;
     case READING:
@@ -460,7 +459,7 @@ CacheIndex::Shutdown()
   }
 
   if (sanitize) {
-    index->RemoveIndexFromDisk();
+    index->RemoveAllIndexFiles();
   }
 
   return NS_OK;
@@ -1356,7 +1355,8 @@ CacheIndex::AsyncGetDiskConsumption(nsICacheStorageConsumptionObserver* aObserve
 
   NS_ENSURE_ARG(observer);
 
-  if (index->mState == READY || index->mState == WRITING) {
+  if ((index->mState == READY || index->mState == WRITING) &&
+      !index->mAsyncGetDiskConsumptionBlocked) {
     LOG(("CacheIndex::AsyncGetDiskConsumption - calling immediately"));
     // Safe to call the callback under the lock,
     // we always post to the main thread.
@@ -1367,6 +1367,20 @@ CacheIndex::AsyncGetDiskConsumption(nsICacheStorageConsumptionObserver* aObserve
   LOG(("CacheIndex::AsyncGetDiskConsumption - remembering callback"));
   // Will be called when the index get to the READY state.
   index->mDiskConsumptionObservers.AppendElement(observer);
+
+  // Move forward with index re/building if it is pending
+  RefPtr<CacheIOThread> ioThread = CacheFileIOManager::IOThread();
+  if (ioThread) {
+    ioThread->Dispatch(NS_NewRunnableFunction([]() -> void {
+      StaticMutexAutoLock lock(sLock);
+
+      RefPtr<CacheIndex> index = gInstance;
+      if (index && index->mUpdateTimer) {
+        index->mUpdateTimer->Cancel();
+        index->DelayedUpdateLocked();
+      }
+    }), CacheIOThread::INDEX);
+  }
 
   return NS_OK;
 }
@@ -1565,7 +1579,7 @@ CacheIndex::ProcessPendingOperations()
 bool
 CacheIndex::WriteIndexToDiskIfNeeded()
 {
-  if (mState != READY || mShuttingDown) {
+  if (mState != READY || mShuttingDown || mRWPending) {
     return false;
   }
 
@@ -1595,6 +1609,7 @@ CacheIndex::WriteIndexToDisk()
   MOZ_ASSERT(mState == READY);
   MOZ_ASSERT(!mRWBuf);
   MOZ_ASSERT(!mRWHash);
+  MOZ_ASSERT(!mRWPending);
 
   ChangeState(WRITING);
 
@@ -1635,6 +1650,7 @@ CacheIndex::WriteRecords()
 
   sLock.AssertCurrentThreadOwns();
   MOZ_ASSERT(mState == WRITING);
+  MOZ_ASSERT(!mRWPending);
 
   int64_t fileOffset;
 
@@ -1711,6 +1727,8 @@ CacheIndex::WriteRecords()
     LOG(("CacheIndex::WriteRecords() - CacheFileIOManager::Write() failed "
          "synchronously [rv=0x%08x]", rv));
     FinishWrite(false);
+  } else {
+    mRWPending = true;
   }
 
   mRWBufPos = 0;
@@ -1724,6 +1742,10 @@ CacheIndex::FinishWrite(bool aSucceeded)
   MOZ_ASSERT((!aSucceeded && mState == SHUTDOWN) || mState == WRITING);
 
   sLock.AssertCurrentThreadOwns();
+
+  // If there is write operation pending we must be cancelling writing of the
+  // index when shutting down or removing the whole index.
+  MOZ_ASSERT(!mRWPending || (!aSucceeded && (mShuttingDown || mRemovingAll)));
 
   mIndexHandle = nullptr;
   mRWHash = nullptr;
@@ -1817,11 +1839,17 @@ CacheIndex::RemoveFile(const nsACString &aName)
 }
 
 void
-CacheIndex::RemoveIndexFromDisk()
+CacheIndex::RemoveAllIndexFiles()
 {
-  LOG(("CacheIndex::RemoveIndexFromDisk()"));
-
+  LOG(("CacheIndex::RemoveAllIndexFiles()"));
   RemoveFile(NS_LITERAL_CSTRING(INDEX_NAME));
+  RemoveJournalAndTempFile();
+}
+
+void
+CacheIndex::RemoveJournalAndTempFile()
+{
+  LOG(("CacheIndex::RemoveJournalAndTempFile()"));
   RemoveFile(NS_LITERAL_CSTRING(TEMP_INDEX_NAME));
   RemoveFile(NS_LITERAL_CSTRING(JOURNAL_NAME));
 }
@@ -1830,8 +1858,7 @@ class WriteLogHelper
 {
 public:
   explicit WriteLogHelper(PRFileDesc *aFD)
-    : mStatus(NS_OK)
-    , mFD(aFD)
+    : mFD(aFD)
     , mBufSize(kMaxBufSize)
     , mBufPos(0)
   {
@@ -1850,11 +1877,10 @@ private:
 
   nsresult FlushBuffer();
 
-  nsresult            mStatus;
-  PRFileDesc         *mFD;
-  char               *mBuf;
-  uint32_t            mBufSize;
-  int32_t             mBufPos;
+  PRFileDesc       *mFD;
+  char             *mBuf;
+  uint32_t          mBufSize;
+  int32_t           mBufPos;
   RefPtr<CacheHash> mHash;
 };
 
@@ -1863,18 +1889,11 @@ WriteLogHelper::AddEntry(CacheIndexEntry *aEntry)
 {
   nsresult rv;
 
-  if (NS_FAILED(mStatus)) {
-    return mStatus;
-  }
-
   if (mBufPos + sizeof(CacheIndexRecord) > mBufSize) {
     mHash->Update(mBuf, mBufPos);
 
     rv = FlushBuffer();
-    if (NS_FAILED(rv)) {
-      mStatus = rv;
-      return rv;
-    }
+    NS_ENSURE_SUCCESS(rv, rv);
     MOZ_ASSERT(mBufPos + sizeof(CacheIndexRecord) <= mBufSize);
   }
 
@@ -1889,17 +1908,10 @@ WriteLogHelper::Finish()
 {
   nsresult rv;
 
-  if (NS_FAILED(mStatus)) {
-    return mStatus;
-  }
-
   mHash->Update(mBuf, mBufPos);
   if (mBufPos + sizeof(CacheHash::Hash32_t) > mBufSize) {
     rv = FlushBuffer();
-    if (NS_FAILED(rv)) {
-      mStatus = rv;
-      return rv;
-    }
+    NS_ENSURE_SUCCESS(rv, rv);
     MOZ_ASSERT(mBufPos + sizeof(CacheHash::Hash32_t) <= mBufSize);
   }
 
@@ -1909,14 +1921,16 @@ WriteLogHelper::Finish()
   rv = FlushBuffer();
   NS_ENSURE_SUCCESS(rv, rv);
 
-  mStatus = NS_ERROR_UNEXPECTED; // Don't allow any other operation
   return NS_OK;
 }
 
 nsresult
 WriteLogHelper::FlushBuffer()
 {
-  MOZ_ASSERT(NS_SUCCEEDED(mStatus));
+  if (CacheObserver::IsPastShutdownIOLag()) {
+    LOG(("WriteLogHelper::FlushBuffer() - Interrupting writing journal."));
+    return NS_ERROR_FAILURE;
+  }
 
   int32_t bytesWritten = PR_Write(mFD, mBuf, mBufPos);
 
@@ -1937,6 +1951,11 @@ CacheIndex::WriteLogToDisk()
 
   MOZ_ASSERT(mPendingUpdates.Count() == 0);
   MOZ_ASSERT(mState == SHUTDOWN);
+
+  if (CacheObserver::IsPastShutdownIOLag()) {
+    LOG(("CacheIndex::WriteLogToDisk() - Skipping writing journal."));
+    return NS_ERROR_FAILURE;
+  }
 
   RemoveFile(NS_LITERAL_CSTRING(TEMP_INDEX_NAME));
 
@@ -1959,9 +1978,11 @@ CacheIndex::WriteLogToDisk()
   for (auto iter = mIndex.Iter(); !iter.Done(); iter.Next()) {
     CacheIndexEntry* entry = iter.Get();
     if (entry->IsRemoved() || entry->IsDirty()) {
-      wlh.AddEntry(entry);
+      rv = wlh.AddEntry(entry);
+      if (NS_WARN_IF(NS_FAILED(rv))) {
+        return rv;
+      }
     }
-    iter.Remove();
   }
 
   rv = wlh.Finish();
@@ -2057,6 +2078,7 @@ CacheIndex::StartReadingIndex()
   MOZ_ASSERT(!mDontMarkIndexClean);
   MOZ_ASSERT(!mJournalReadSuccessfully);
   MOZ_ASSERT(mIndexHandle->FileSize() >= 0);
+  MOZ_ASSERT(!mRWPending);
 
   int64_t entriesSize = mIndexHandle->FileSize() - sizeof(CacheIndexHeader) -
                         sizeof(CacheHash::Hash32_t);
@@ -2079,6 +2101,8 @@ CacheIndex::StartReadingIndex()
     LOG(("CacheIndex::StartReadingIndex() - CacheFileIOManager::Read() failed "
          "synchronously [rv=0x%08x]", rv));
     FinishRead(false);
+  } else {
+    mRWPending = true;
   }
 }
 
@@ -2090,6 +2114,8 @@ CacheIndex::ParseRecords()
   nsresult rv;
 
   sLock.AssertCurrentThreadOwns();
+
+  MOZ_ASSERT(!mRWPending);
 
   uint32_t entryCnt = (mIndexHandle->FileSize() - sizeof(CacheIndexHeader) -
                      sizeof(CacheHash::Hash32_t)) / sizeof(CacheIndexRecord);
@@ -2174,10 +2200,10 @@ CacheIndex::ParseRecords()
 
   MOZ_ASSERT(fileOffset <= mIndexHandle->FileSize());
   if (fileOffset == mIndexHandle->FileSize()) {
-    if (mRWHash->GetHash() != NetworkEndian::readUint32(mRWBuf)) {
+    uint32_t expectedHash = NetworkEndian::readUint32(mRWBuf);
+    if (mRWHash->GetHash() != expectedHash) {
       LOG(("CacheIndex::ParseRecords() - Hash mismatch, [is %x, should be %x]",
-           mRWHash->GetHash(),
-           NetworkEndian::readUint32(mRWBuf)));
+           mRWHash->GetHash(), expectedHash));
       FinishRead(false);
       return;
     }
@@ -2207,6 +2233,8 @@ CacheIndex::ParseRecords()
          "synchronously [rv=0x%08x]", rv));
     FinishRead(false);
     return;
+  } else {
+    mRWPending = true;
   }
 }
 
@@ -2223,6 +2251,7 @@ CacheIndex::StartReadingJournal()
   MOZ_ASSERT(mIndexOnDiskIsValid);
   MOZ_ASSERT(mTmpJournal.Count() == 0);
   MOZ_ASSERT(mJournalHandle->FileSize() >= 0);
+  MOZ_ASSERT(!mRWPending);
 
   int64_t entriesSize = mJournalHandle->FileSize() -
                         sizeof(CacheHash::Hash32_t);
@@ -2244,6 +2273,8 @@ CacheIndex::StartReadingJournal()
     LOG(("CacheIndex::StartReadingJournal() - CacheFileIOManager::Read() failed"
          " synchronously [rv=0x%08x]", rv));
     FinishRead(false);
+  } else {
+    mRWPending = true;
   }
 }
 
@@ -2255,6 +2286,8 @@ CacheIndex::ParseJournal()
   nsresult rv;
 
   sLock.AssertCurrentThreadOwns();
+
+  MOZ_ASSERT(!mRWPending);
 
   uint32_t entryCnt = (mJournalHandle->FileSize() -
                        sizeof(CacheHash::Hash32_t)) / sizeof(CacheIndexRecord);
@@ -2295,10 +2328,10 @@ CacheIndex::ParseJournal()
 
   MOZ_ASSERT(fileOffset <= mJournalHandle->FileSize());
   if (fileOffset == mJournalHandle->FileSize()) {
-    if (mRWHash->GetHash() != NetworkEndian::readUint32(mRWBuf)) {
+    uint32_t expectedHash = NetworkEndian::readUint32(mRWBuf);
+    if (mRWHash->GetHash() != expectedHash) {
       LOG(("CacheIndex::ParseJournal() - Hash mismatch, [is %x, should be %x]",
-           mRWHash->GetHash(),
-           NetworkEndian::readUint32(mRWBuf)));
+           mRWHash->GetHash(), expectedHash));
       FinishRead(false);
       return;
     }
@@ -2321,6 +2354,8 @@ CacheIndex::ParseJournal()
          "synchronously [rv=0x%08x]", rv));
     FinishRead(false);
     return;
+  } else {
+    mRWPending = true;
   }
 }
 
@@ -2405,6 +2440,10 @@ CacheIndex::FinishRead(bool aSucceeded)
     // -> ready
     (aSucceeded && mIndexOnDiskIsValid && mJournalReadSuccessfully));
 
+  // If there is read operation pending we must be cancelling reading of the
+  // index when shutting down or removing the whole index.
+  MOZ_ASSERT(!mRWPending || (!aSucceeded && (mShuttingDown || mRemovingAll)));
+
   if (mState == SHUTDOWN) {
     RemoveFile(NS_LITERAL_CSTRING(TEMP_INDEX_NAME));
     RemoveFile(NS_LITERAL_CSTRING(JOURNAL_NAME));
@@ -2473,30 +2512,41 @@ CacheIndex::DelayedUpdate(nsITimer *aTimer, void *aClosure)
 {
   LOG(("CacheIndex::DelayedUpdate()"));
 
-  nsresult rv;
   StaticMutexAutoLock lock(sLock);
-
   RefPtr<CacheIndex> index = gInstance;
 
   if (!index) {
     return;
   }
 
-  index->mUpdateTimer = nullptr;
+  index->DelayedUpdateLocked();
+}
 
-  if (!index->IsIndexUsable()) {
+// static
+void
+CacheIndex::DelayedUpdateLocked()
+{
+  LOG(("CacheIndex::DelayedUpdateLocked()"));
+
+  sLock.AssertCurrentThreadOwns();
+
+  nsresult rv;
+
+  mUpdateTimer = nullptr;
+
+  if (!IsIndexUsable()) {
     return;
   }
 
-  if (index->mState == READY && index->mShuttingDown) {
+  if (mState == READY && mShuttingDown) {
     return;
   }
 
   // mUpdateEventPending must be false here since StartUpdatingIndex() won't
   // schedule timer if it is true.
-  MOZ_ASSERT(!index->mUpdateEventPending);
-  if (index->mState != BUILDING && index->mState != UPDATING) {
-    LOG(("CacheIndex::DelayedUpdate() - Update was canceled"));
+  MOZ_ASSERT(!mUpdateEventPending);
+  if (mState != BUILDING && mState != UPDATING) {
+    LOG(("CacheIndex::DelayedUpdateLocked() - Update was canceled"));
     return;
   }
 
@@ -2504,13 +2554,13 @@ CacheIndex::DelayedUpdate(nsITimer *aTimer, void *aClosure)
   RefPtr<CacheIOThread> ioThread = CacheFileIOManager::IOThread();
   MOZ_ASSERT(ioThread);
 
-  index->mUpdateEventPending = true;
-  rv = ioThread->Dispatch(index, CacheIOThread::INDEX);
+  mUpdateEventPending = true;
+  rv = ioThread->Dispatch(this, CacheIOThread::INDEX);
   if (NS_FAILED(rv)) {
-    index->mUpdateEventPending = false;
-    NS_WARNING("CacheIndex::DelayedUpdate() - Can't dispatch event");
+    mUpdateEventPending = false;
+    NS_WARNING("CacheIndex::DelayedUpdateLocked() - Can't dispatch event");
     LOG(("CacheIndex::DelayedUpdate() - Can't dispatch event" ));
-    index->FinishUpdate(false);
+    FinishUpdate(false);
   }
 }
 
@@ -3125,7 +3175,13 @@ CacheIndex::ChangeState(EState aNewState)
     CacheFileIOManager::CacheIndexStateChanged();
   }
 
-  if (mState == READY && mDiskConsumptionObservers.Length()) {
+  NotifyAsyncGetDiskConsumptionCallbacks();
+}
+
+void
+CacheIndex::NotifyAsyncGetDiskConsumptionCallbacks()
+{
+  if ((mState == READY || mState == WRITING) && !mAsyncGetDiskConsumptionBlocked && mDiskConsumptionObservers.Length()) {
     for (uint32_t i = 0; i < mDiskConsumptionObservers.Length(); ++i) {
       DiskConsumptionObserver* o = mDiskConsumptionObservers[i];
       // Safe to call under the lock.  We always post to the main thread.
@@ -3160,9 +3216,13 @@ CacheIndex::AllocBuffer()
 void
 CacheIndex::ReleaseBuffer()
 {
-  if (!mRWBuf) {
+  sLock.AssertCurrentThreadOwns();
+
+  if (!mRWBuf || mRWPending) {
     return;
   }
+
+  LOG(("CacheIndex::ReleaseBuffer() releasing buffer"));
 
   free(mRWBuf);
   mRWBuf = nullptr;
@@ -3271,13 +3331,13 @@ CacheIndex::OnFileOpenedInternal(FileOpenHelper *aOpener,
   LOG(("CacheIndex::OnFileOpenedInternal() [opener=%p, handle=%p, "
        "result=0x%08x]", aOpener, aHandle, aResult));
 
+  MOZ_ASSERT(CacheFileIOManager::IsOnIOThread());
+
   nsresult rv;
 
   sLock.AssertCurrentThreadOwns();
 
-  if (!IsIndexUsable()) {
-    return NS_ERROR_NOT_AVAILABLE;
-  }
+  MOZ_RELEASE_ASSERT(IsIndexUsable());
 
   if (mState == READY && mShuttingDown) {
     return NS_OK;
@@ -3381,13 +3441,15 @@ CacheIndex::OnDataWritten(CacheFileHandle *aHandle, const char *aBuf,
   LOG(("CacheIndex::OnDataWritten() [handle=%p, result=0x%08x]", aHandle,
        aResult));
 
+  MOZ_ASSERT(CacheFileIOManager::IsOnIOThread());
+
   nsresult rv;
 
   StaticMutexAutoLock lock(sLock);
 
-  if (!IsIndexUsable()) {
-    return NS_ERROR_NOT_AVAILABLE;
-  }
+  MOZ_RELEASE_ASSERT(IsIndexUsable());
+  MOZ_RELEASE_ASSERT(mRWPending);
+  mRWPending = false;
 
   if (mState == READY && mShuttingDown) {
     return NS_OK;
@@ -3395,11 +3457,7 @@ CacheIndex::OnDataWritten(CacheFileHandle *aHandle, const char *aBuf,
 
   switch (mState) {
     case WRITING:
-      if (mIndexHandle != aHandle) {
-        LOG(("CacheIndex::OnDataWritten() - ignoring notification since it "
-             "belongs to previously canceled operation [state=%d]", mState));
-        break;
-      }
+      MOZ_ASSERT(mIndexHandle == aHandle);
 
       if (NS_FAILED(aResult)) {
         FinishWrite(false);
@@ -3422,6 +3480,7 @@ CacheIndex::OnDataWritten(CacheFileHandle *aHandle, const char *aBuf,
       // Writing was canceled.
       LOG(("CacheIndex::OnDataWritten() - ignoring notification since the "
            "operation was previously canceled [state=%d]", mState));
+      ReleaseBuffer();
   }
 
   return NS_OK;
@@ -3433,11 +3492,13 @@ CacheIndex::OnDataRead(CacheFileHandle *aHandle, char *aBuf, nsresult aResult)
   LOG(("CacheIndex::OnDataRead() [handle=%p, result=0x%08x]", aHandle,
        aResult));
 
+  MOZ_ASSERT(CacheFileIOManager::IsOnIOThread());
+
   StaticMutexAutoLock lock(sLock);
 
-  if (!IsIndexUsable()) {
-    return NS_ERROR_NOT_AVAILABLE;
-  }
+  MOZ_RELEASE_ASSERT(IsIndexUsable());
+  MOZ_RELEASE_ASSERT(mRWPending);
+  mRWPending = false;
 
   switch (mState) {
     case READING:
@@ -3457,6 +3518,7 @@ CacheIndex::OnDataRead(CacheFileHandle *aHandle, char *aBuf, nsresult aResult)
       // Reading was canceled.
       LOG(("CacheIndex::OnDataRead() - ignoring notification since the "
            "operation was previously canceled [state=%d]", mState));
+      ReleaseBuffer();
   }
 
   return NS_OK;
@@ -3482,11 +3544,11 @@ CacheIndex::OnFileRenamed(CacheFileHandle *aHandle, nsresult aResult)
   LOG(("CacheIndex::OnFileRenamed() [handle=%p, result=0x%08x]", aHandle,
        aResult));
 
+  MOZ_ASSERT(CacheFileIOManager::IsOnIOThread());
+
   StaticMutexAutoLock lock(sLock);
 
-  if (!IsIndexUsable()) {
-    return NS_ERROR_NOT_AVAILABLE;
-  }
+  MOZ_RELEASE_ASSERT(IsIndexUsable());
 
   if (mState == READY && mShuttingDown) {
     return NS_OK;
@@ -3657,6 +3719,22 @@ CacheIndex::ReportHashStats()
   }
 
   CacheObserver::SetHashStatsReported();
+}
+
+// static
+void
+CacheIndex::OnAsyncEviction(bool aEvicting)
+{
+  RefPtr<CacheIndex> index = gInstance;
+  if (!index) {
+    return;
+  }
+
+  StaticMutexAutoLock lock(sLock);
+  index->mAsyncGetDiskConsumptionBlocked = aEvicting;
+  if (!aEvicting) {
+    index->NotifyAsyncGetDiskConsumptionCallbacks();
+  }
 }
 
 } // namespace net
